Add transfer-learning support and fine-tuning
Add transfer-learning workflow, hardware accel and fine-tuning support. app.py: use CNN model as fallback when a selected transfer model is missing, ensure uploaded images are saved as RGB JPEGs, and adjust st.image sizing. src/model.py: force output Dense layers to float32 to avoid mixed-precision issues. src/train.py: import transfer model, add mixed-precision/XLA and memory-growth configuration, implement multi-stage training with optional fine-tuning (unfreeze last N layers), history merging, and model_type selection ('cnn' or 'transfer'). train_model.bat: add model_type and fine_tune_epochs arguments and pass them to the trainer. Also add a generated model history JSON file.

+++ src/model.py
@@ -67,7 +67,7 @@ def create_cnn_model(input_shape=(224, 224, 3), num_classes=2):
         layers.Dropout(0.5),
         
         # Output Layer
-        layers.Dense(num_classes, activation='softmax')
+        layers.Dense(num_classes, activation='softmax', dtype='float32')
     ])
     
     return model
@@ -100,7 +100,7 @@ def create_transfer_learning_model(input_shape=(224, 224, 3), num_classes=2):
         layers.GlobalAveragePooling2D(),
         layers.Dense(256, activation='relu'),
         layers.Dropout(0.5),
-        layers.Dense(num_classes, activation='softmax')
+        layers.Dense(num_classes, activation='softmax', dtype='float32')
     ])
     
     return model
